+++ Westwind.MessageQueueing.WebHostSample/_Code/oWinStartup.cs
@@ -1,6 +1,4 @@
-﻿using Microsoft.AspNet.SignalR;
-using Microsoft.Owin;
-using Microsoft.Owin.Cors;
+﻿using Microsoft.Owin;
 using Owin;
 using QueueManagerStarter;